pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "federated-skin-cancer"
version = "0.1.0"
description = "Federated Learning for Skin Cancer Classification with LMS-ViT"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Leonidas Brando", email = "leonidas.brando@alumnos.upm.es"}
]
keywords = [
    "federated-learning",
    "skin-cancer",
    "vision-transformer",
    "medical-imaging",
    "deep-learning"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "timm>=0.9.12",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "Pillow>=10.0.0",
    "scikit-learn>=1.3.0",
    "PyYAML>=6.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]
notebooks = [
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
]
monitoring = [
    "tensorboard>=2.14.0",
    "wandb>=0.15.0",
]

[project.urls]
Homepage = "https://github.com/leonidasdev/federated-light-skin-cancer-classification"
Repository = "https://github.com/leonidasdev/federated-light-skin-cancer-classification"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ['py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | build
    | dist
    | data
    | checkpoints
)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip = [".git", ".venv", "build", "dist"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"
